refactor(core): split storage service responsibilities
