+++ framework/core/architecture/LAYERS.md
@@ -42,7 +42,7 @@ remain private implementation details behind the public `kungfu` facade.
 | `yijinjing-schema` | `schema-values` | `core/schema` | 5 | `yijinjing` | `src/libyijinjing/include/kungfu/yijinjing/schema/core.h` |
 | `yijinjing-kernel` | `journal-kernel` | `core/yijinjing` | 42 | `yijinjing` | `src/libyijinjing/include/kungfu/yijinjing/journal/journal.h`<br>`src/libyijinjing/include/kungfu/yijinjing/storage.h` |
 | `libkungfu-contracts` | `ports-contracts` | `core/runtime-contracts` | 55 | `yijinjing`<br>`kungfu_contracts`<br>`kungfu` | `src/libkungfu/include/kungfu/runtime/common.h`<br>`src/libkungfu/include/kungfu/runtime/storage/service.h` |
-| `libkungfu-services` | `application-services` | `core/runtime-services` | 33 | `kungfu_services`<br>`kungfu_services_state_cache`<br>`kungfu` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/state_service.cpp`<br>`src/libkungfu/src/runtime/query/fact_query.cpp` |
+| `libkungfu-services` | `application-services` | `core/runtime-services` | 36 | `kungfu_services`<br>`kungfu_services_state_cache`<br>`kungfu` | `src/libkungfu/src/runtime/storage/service.cpp`<br>`src/libkungfu/src/runtime/state_service.cpp`<br>`src/libkungfu/src/runtime/query/fact_query.cpp` |
 | `libkungfu-adapters` | `adapters` | `core/runtime-adapters` | 14 | `kungfu_adapters`<br>`kungfu`<br>`kungfu_native_storage_shared` | `src/libkungfu/src/view/schema.cpp`<br>`src/libkungfu/src/runtime/native_storage.cpp`<br>`src/libkungfu/src/runtime/util/rocks.cpp` |
 | `core-composition-bindings` | `composition-bindings` | `core/bindings` | 43 | `kungfu_composition`<br>`kungfu`<br>`kungfu_embedding`<br>`kungfu_wasm_host`<br>`kungfu_node`<br>`kungfu_electron`<br>`drone`<br>`kungfu_kfc`<br>`kungfu_node_host`<br>`pykungfu` | `src/bindings/node/binding/kungfu_node.cpp`<br>`src/bindings/python/binding/pykungfu.cpp`<br>`src/libkungfu/src/runtime/embedding.cpp` |
 | `core-native-qualification` | `qualification` | `core/qualification` | 16 | `yijinjing_mmap_tests`<br>`yijinjing_content_hash_tests`<br>`yijinjing_mmap_qualification`<br>`kungfu_durability_contract_tests`<br>`kungfu_runtime_error_tests`<br>`kungfu_embedding_generic_codec_tests`<br>`kungfu_peer_continuity_tests`<br>`kungfu_state_service_contract_tests`<br>`kungfu_durable_ingest_tests`<br>`kungfu_durability_powercut_fixture`<br>`kungfu_durability_slo_fixture`<br>`kungfu_offhost_backup_fixture`<br>`kungfu_projection_bootstrap_tests`<br>`kungfu_crash_recovery_tests`<br>`kungfu_profile_lifecycle_tests` | `src/libkungfu/tests/durability_contract_tests.cpp`<br>`src/libyijinjing/tests/mmap_tests.cpp` |
@@ -56,7 +56,7 @@ from the same authority as this map.
 | Target | Kind | Component | Depends on | Sources |
 | --- | --- | --- | --- | ---: |
 | `kungfu_contracts` | `INTERFACE` | `libkungfu-contracts` | — | 0 |
-| `kungfu_services` | `OBJECT` | `libkungfu-services` | `kungfu_contracts` | 29 |
+| `kungfu_services` | `OBJECT` | `libkungfu-services` | `kungfu_contracts` | 32 |
 | `kungfu_services_state_cache` | `OBJECT` | `libkungfu-services` | `kungfu_contracts` | 3 |
 | `kungfu_adapters` | `OBJECT` | `libkungfu-adapters` | `kungfu_services`<br>`kungfu_services_state_cache` | 13 |
 | `kungfu_composition` | `OBJECT` | `core-composition-bindings` | `kungfu_adapters` | 3 |
@@ -68,7 +68,10 @@ back into the compatibility facade.
 
 | Responsibility | Source | Line budget |
 | --- | --- | ---: |
-| Public storage facade, shared application helpers and compatibility surface | `src/libkungfu/src/runtime/storage/service.cpp` | 5000 |
+| Public storage facade and shared application-service composition | `src/libkungfu/src/runtime/storage/service.cpp` | 3000 |
+| Episode repair planning, evidence fetch, bundle validation and non-destructive apply | `src/libkungfu/src/runtime/storage/episode_repair.cpp` | 1500 |
+| Typed journal queries and stable storage query result rendering | `src/libkungfu/src/runtime/storage/query_render.cpp` | 500 |
+| JSON compatibility decoding, validation and typed option parsing | `src/libkungfu/src/runtime/storage/json_compat.cpp` | 500 |
 | File and RocksDB content providers behind the storage port | `src/libkungfu/src/runtime/storage/provider.cpp` | 800 |
 | Status, fsck, projection rebuild, GC and compaction planning | `src/libkungfu/src/runtime/storage/maintenance_service.cpp` | 500 |
 | Manifest bundle import, export and sync verification | `src/libkungfu/src/runtime/storage/transfer_service.cpp` | 520 |
