+++ framework/core/architecture/TARGETS.cmake
@@ -28,8 +28,11 @@ set(KUNGFU_SERVICES_SOURCE_FILES
   "${PROJECT_SOURCE_DIR}/src/runtime/state_service.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/state_shadow.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/episode_manifest_projection.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/episode_repair.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/json_compat.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/maintenance_service.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/manifest_catalog_projection.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/query_render.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/service.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/source_registry_projection.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/transfer_service.cpp"
